// src/app/models/admin/dangkyhoatdongkhoangsan/dangkykhaithac/dkkhaithacloaikhoangsan.model.ts
export class OutputDkKhaiThacLoaiKhoangSan {
  idkhaithacloaikhoangsan: string;
  idloaikhoangsan: string;
  tenloaikhoangsan: string;
  iddangkykhaithac: string;
  tenkhoangsan: string;
  truluong: number;
  donvitruluong: string;
  siteid: string;
}

export class InputDkKhaiThacLoaiKhoangSan {
  idkhaithacloaikhoangsan: string;
  idloaikhoangsan: string;
  iddangkykhaithac: string;
  tenkhoangsan: string;
  truluong: number;
  donvitruluong: string;
  siteid: string;
}